// libs/components/forms/cascade-select/src/lib/cascade-select.types.ts
export type DynamoCascadeSelectPart =
  | 'root'
  | 'trigger'
  | 'chevron'
  | 'clear'
  | 'panel'
  | 'row'
  | 'caret'
  | 'checkbox'
  | 'filterInput'
  | 'no-results';
export type DynamoCascadeSelectSelectionMode =
  'single' | 'multiple' | 'checkbox';
export type DynamoCascadePosition =
  'right-start' | 'right-end' | 'left-start' | 'left-end';